Responses

This module is responsible for serializing returned values from route handlers into valid HTTP.Responses. Any value returned from a handler is supported. Strings are content-sniffed using HTTP.sniff to determine Content-Type. Base.Docs.HTML and Base.Docs.Text are set to text/html and text/plain respectively. Simple values such as numbers, symbols, and characters are sent as text/plain. Other values are serialized as application/json using the JSON package, using the allownan = true keyword.

The module exports a response helper function that allows for declaring the intended serialization mime type, along with other options such as attachment and filename which effect the Content-Disposition header.

ReloadableMiddleware.Responses.response — Method

response(mime, object; headers = nothing, attachment = false, filename = nothing)

Construct a response object from object in the provided mime type and set the Content-Disposition based on filename and attachment. Will automatically set Content-Type and Content-Length based on the object and mime given. mime can either be a MIME, or a String.

When object is a String or Vector{UInt8} then no transformation is applied to the object and it is instead used directly in the Response object. Any other type is first rendered using show(io, mime, object) to create the correct mime type output.

Additional response headers can be passed in the headers keyword. Any iterable that returns k/v pairs can be used as the argument.

ReloadableMiddleware.Responses.sse — Function

sse(stream, [data]; [event,] [id,] [retry])

Format a server sent event. Write it to stream if provided. event, id, and retry are optional. When data is not provided then a single line data: \n is written to the stream.